Fix: 앨범 수정, 삭제 오류 수정
